# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Vehicle:
    properties:
      color:
        type: string
      plateNumber:
        type: string
      userID:
        type: integer
      vehicleID:
        type: integer
    type: object
  models.ViolationRecord:
    properties:
      fineAmount:
        type: number
      recordID:
        type: integer
      status:
        type: string
      violationID:
        type: integer
      violationType:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: 1351688749@qq.com
    name: API Support
    url: http://example.com/support
  description: API documentation for Parking Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Parking Management API
  version: "1.0"
paths:
...
  /api/vehicle:
    post:
      consumes:
      - application/json
      description: Create a new vehicle
      parameters:
      - description: Vehicle information
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/models.Vehicle'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vehicle'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new vehicle
      tags:
      - Vehicle
  /api/vehicle/{id}:
    delete:
      consumes:
      - application/json
      description: Delete vehicle by ID
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete vehicle by ID
      tags:
      - Vehicle
    get:
      consumes:
      - application/json
      description: Get vehicle by ID
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vehicle'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get vehicle by ID
      tags:
      - Vehicle
    put:
      consumes:
      - application/json
      description: Update vehicle by ID
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Vehicle information
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/models.Vehicle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vehicle'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update vehicle by ID
      tags:
      - Vehicle
  /api/vehicle/user/{id}:
    get:
      consumes:
      - application/json
      description: Get all vehicles by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vehicle'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all vehicles by user ID
      tags:
      - Vehicle
  /api/violationrecord:
    post:
      consumes:
      - application/json
      description: Create a new violation record
      parameters:
      - description: Violation record information
        in: body
        name: violationRecord
        required: true
        schema:
          $ref: '#/definitions/models.ViolationRecord'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ViolationRecord'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new violation record
      tags:
      - ViolationRecord
  /api/violationrecord/fineamount/record/{id}:
    get:
      consumes:
      - application/json
      description: Get fine amount by record ID
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ViolationRecord'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get fine amount by record ID
      tags:
      - ViolationRecord
  /api/violationrecord/status/record/{id}:
    get:
      consumes:
      - application/json
      description: Get status by record ID
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ViolationRecord'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get status by record ID
      tags:
      - ViolationRecord
  /api/violationrecord/type/record/{id}:
    get:
      consumes:
      - application/json
      description: Get violation type by record ID
      parameters:
      - description: Record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ViolationRecord'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get violation type by record ID
      tags:
      - ViolationRecord
  /api/violationrecord/user/{id}:
    get:
      consumes:
      - application/json
      description: Get violation records by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ViolationRecord'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get violation records by user ID
      tags:
      - ViolationRecord
  /api/violationrecord/violation/{type}:
    get:
      consumes:
      - application/json
      description: Statistical violations by type
      parameters:
      - description: Violation type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ViolationRecord'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Statistical violations by type
      tags:
      - ViolationRecord
...
swagger: "2.0"